Revert "impl package_document"
