--- src/epub/mod.rs
@@ -1,5 +1,0 @@
-mod package_document;
-mod media_type;
-
-pub use package_document::{PackageDocument, Dir, meta_data::*, manifest::*, spine::*};
-pub use media_type::*;
