+++ src/lib.rs
@@ -1,20 +1,160 @@
-pub mod read;
-pub mod epub;
+mod media_type;
 
-pub mod prelude {
-    pub use crate::EPUBError;
-    pub use crate::read::EPUBReader;
-    pub use crate::epub::*;
+use std::io::{Read, Seek, BufReader};
+use std::fs::File;
+use std::path::{Path, PathBuf};
+use zip::read::{ZipArchive, ZipFile};
+use failure::{Fail, Error};
+use failure::_core::ops::Deref;
+
+#[derive(Debug)]
+pub struct EPUBReader<R: Read + Seek> {
+    archive: ZipArchive<R>,
 }
 
-use failure::Fail;
+impl<R: Read + Seek> EPUBReader<R> {
+    const MIMETYPE_PATH: &'static str = "mimetype";
+    const MIMETYPE_CONTENT: &'static str = "application/epub+zip";
+    const CONTAINER_XML_PATH: &'static str = "META-INF/container.xml";
+}
+
+impl EPUBReader<BufReader<File>> {
+    pub fn new<P>(path: P) -> Result<Self, Error> where P: AsRef<Path> {
+        let archive: ZipArchive<BufReader<File>> = std::fs::File::open(path)
+            .map(|file| {
+                let buf_reader = std::io::BufReader::new(file);
+                ZipArchive::new(buf_reader)
+            })??;
+
+        Ok(Self {
+            archive
+        })
+    }
+
+
+    fn package_document_paths(&mut self) -> Result<Vec<PathBuf>, Error> {
+        use xml::attribute::OwnedAttribute;
+        use xml::reader::XmlEvent;
+
+        let archive = &mut self.archive;
+        let container_xml = archive.by_name(Self::CONTAINER_XML_PATH)?;
+
+        let file = BufReader::new(container_xml);
+        let mut parser = xml::EventReader::new(file);
+
+        let package_documents = parser.into_iter()
+            .filter_map(|e| match e.ok()? {
+                XmlEvent::StartElement { name, attributes, .. } => {
+                    if name.local_name == "rootfile".to_string() {
+                        Some(
+                            attributes.into_iter()
+                                .find_map(|atr| {
+                                    let OwnedAttribute { name, value } = atr;
+                                    if name.local_name == "full-path".to_string() {
+                                        let path_buf = PathBuf::from(value);
+                                        Some(path_buf)
+                                    } else {
+                                        None
+                                    }
+                                })?
+                        )
+                    } else {
+                        None
+                    }
+                }
+                _ => None
+            })
+            .collect::<Vec<PathBuf>>();
+
+        Ok(package_documents)
+    }
+
+    fn package_document_path(&mut self) -> Result<Option<PathBuf>, Error> {
+        self.package_document_paths()
+            .map(|ps| ps.into_iter().next())
+    }
+
+    pub fn check_mimetype(&mut self) -> Result<(), Error> {
+        let archive = &mut self.archive;
+        let mut mimetype = archive.by_index(0)?;
+        if mimetype.name() != Self::MIMETYPE_PATH {
+            Err(EPUBError::MimeTypeError {
+                err_msg: "The mimetype not found".to_string()
+            })?
+        }
+        let mut mimetype_content = [0u8; 20];
+        mimetype.read(&mut mimetype_content);
+        match String::from_utf8(mimetype_content.to_vec()) {
+            Ok(content) if content == Self::MIMETYPE_CONTENT.to_string() => {}
+            Ok(content) => {
+                Err(EPUBError::MimeTypeError {
+                    err_msg: format!("The mimetype file's content {} is invalid. It MUST be `application/epub+zip`.", &content)
+                })?
+            }
+            Err(e) => {
+                Err(EPUBError::MimeTypeError {
+                    err_msg: format!("{}", std::error::Error::description(&e))
+                })?
+            }
+        }
+
+        Ok(())
+    }
+
+    pub fn check_container_xml(&mut self) -> Result<(), Error> {
+        use xml::name::OwnedName;
+        use xml::attribute::OwnedAttribute;
+        use xml::reader::XmlEvent;
+
+        let archive = &mut self.archive;
+        let container_xml = archive.by_name(Self::CONTAINER_XML_PATH)?;
+
+        let file = BufReader::new(container_xml);
+        let mut parser = xml::EventReader::new(file);
+
+        parser.into_iter()
+            .filter_map(|e| match e.ok()? {
+                XmlEvent::StartElement { name, attributes, .. } => {
+                    if name.local_name == "rootfile".to_string() {
+                        Some(
+                            attributes.into_iter()
+                                .find_map(|atr| {
+                                    let OwnedAttribute { name, value } = atr;
+                                    if name.local_name == "full-path".to_string() {
+                                        let path_buf = PathBuf::from(value);
+                                        Some(path_buf)
+                                    } else {
+                                        None
+                                    }
+                                })
+                                .ok_or(
+                                    EPUBError::ContainerError {
+                                        err_msg: "Full-path of package document is undefined in <rootfile>".to_string()
+                                    }.into()
+                                )
+                        )
+                    } else {
+                        None
+                    }
+                }
+                _ => None
+            })
+            .collect::<Result<Vec<PathBuf>, Error>>()?;
+
+        Ok(())
+    }
+}
+
+impl<R: Read + Seek> Deref for EPUBReader<R> {
+    type Target = ZipArchive<R>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.archive
+    }
+}
 
 #[derive(Fail, Debug)]
 pub enum EPUBError {
-    #[fail(display = "PACKAGE DOCUMENT :ERROR\t: {}", err_msg)]
-    PackageDocumentError {
-        err_msg: String,
-    },
     #[fail(display = "MEDIA TYPE :ERROR\t: {}", err_msg)]
     MediaTypeError {
         err_msg: String,
@@ -37,151 +177,10 @@ pub enum EPUBError {
     },
 }
 
-pub mod util {
-    use std::io::Read;
-    use std::iter::Peekable;
-    use xml::reader::*;
-    use failure::_core::ops::Deref;
-
-    #[derive(Debug, Clone)]
-    pub struct Xml {
-        pub vec: Vec<XmlElement>
-    }
-
-    impl Deref for Xml {
-        type Target = Vec<XmlElement>;
-
-        fn deref(&self) -> &Self::Target {
-            &self.vec
-        }
-    }
-
-    impl Xml {
-        pub fn new<R: Read>(iter: &mut Peekable<Events<R>>) -> Self {
-            let mut vec = Vec::new();
-
-            while let Some(_) = iter.peek() {
-                if let Some(elem) = XmlElement::new(iter) {
-                    vec.push(elem);
-                }
-            }
-
-            Self {
-                vec
-            }
-        }
-
-        pub fn get_by<F: FnMut(&XmlElement) -> bool>(&self, mut f: F) -> Option<&XmlElement> {
-            self.vec.iter()
-                .find_map(|e| e.get_by(&mut f))
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct XmlElement {
-        pub event: XmlEvent,
-        pub children: Vec<XmlElement>,
-    }
-
-    impl XmlElement {
-        pub fn new<R: Read>(iter: &mut Peekable<Events<R>>) -> Option<Self> {
-            let mut children = Vec::new();
-
-            match iter.peek()?.as_ref().ok()? {
-                XmlEvent::CData(_)
-                | XmlEvent::Characters(_)
-                | XmlEvent::Comment(_)
-                | XmlEvent::Whitespace(_)
-                | XmlEvent::ProcessingInstruction { .. } => {
-                    let event = iter.next()?.ok()?;
-
-                    Some(Self {
-                        event,
-                        children,
-                    })
-                }
-                XmlEvent::StartElement { name, .. } => {
-                    let start_name = name.clone();
-                    let event = iter.next()?.ok()?;
-
-                    while let Some(child) = Self::new(iter) {
-                        children.push(child);
-
-                        match iter.peek()?.as_ref().ok()? {
-                            XmlEvent::EndElement { name } if &start_name == name => {
-                                let _ = iter.next();
-                                return Some(Self {
-                                    event,
-                                    children,
-                                });
-                            }
-                            _ => {}
-                        }
-                    }
-
-                    Some(Self {
-                        event,
-                        children,
-                    })
-                }
-                XmlEvent::StartDocument { .. } => {
-                    let event = iter.next()?.ok()?;
-
-                    while let Some(child) = Self::new(iter) {
-                        children.push(child);
-
-                        match iter.peek()?.as_ref().ok()? {
-                            XmlEvent::EndDocument => {
-                                let _ = iter.next();
-                                return Some(Self {
-                                    event,
-                                    children,
-                                });
-                            }
-                            _ => {}
-                        }
-                    }
-
-                    Some(Self {
-                        event,
-                        children,
-                    })
-                }
-                _ => {
-                    let _ = iter.next();
-                    None
-                }
-            }
-        }
-
-        pub fn inner_text(&self) -> String {
-            self.children.iter()
-                .filter_map(|e| {
-                    match &e.event {
-                        XmlEvent::CData(cdata) => Some(cdata.to_string()),
-                        XmlEvent::Characters(characters) => Some(characters.to_string()),
-                        XmlEvent::StartDocument { .. } | XmlEvent::StartElement { .. } => {
-                            Some(e.inner_text())
-                        }
-                        XmlEvent::EndDocument
-                        | XmlEvent::EndElement { .. }
-                        | XmlEvent::Whitespace(..)
-                        | XmlEvent::ProcessingInstruction { .. }
-                        | XmlEvent::Comment(..) => {
-                            None
-                        }
-                    }
-                })
-                .fold(String::new(), |acc, s| acc + &s)
-        }
-
-        pub fn get_by<F: FnMut(&XmlElement) -> bool>(&self, f: &mut F) -> Option<&XmlElement> {
-            if f(self) {
-                Some(&self)
-            } else {
-                self.children.iter()
-                    .find(|&e| f(e))
-            }
-        }
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn it_works() {
+        assert_eq!(2 + 2, 4);
     }
 }
