+++ src/main.rs
@@ -1,6 +1,9 @@
-use epub_rs::prelude::*;
+use epub_rs::*;
 
 fn main() -> Result<(), failure::Error>{
+    use std::io::{Read, BufReader};
+    use zip::ZipArchive;
+
     let path = "tests/data/childrens-literature.epub";
     // let file = std::fs::File::open(path)?;
     // let buf_reader = BufReader::new(file);
@@ -10,10 +13,7 @@ fn main() -> Result<(), failure::Error>{
     //     .inspect(|n| println!("{}", n))
     //     .last();
 
-    let epub = EPUBReader::new(path)?;
+    let mut epub = EPUBReader::new(path)?;
     // println!("{:?}", epub.package_document_paths());
-    for package_document in epub.package_documents {
-        println!("{:?}", package_document);
-    }
     Ok(())
 }
