--- tests/data/sample_package.opf
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package unique-identifier="idName" version="3.0" xmlns="http://www.idpf.org/2007/opf" xml:lang="ja">
-    <!-- 書籍情報 -->
-    <metadata xmlns:dc="http://purl.org/dc/elements/1.1/">
-        <dc:identifier id="idName">urn:uuid:00000000-0000-0000-0000-000000000000</dc:identifier>
-        <dc:publisher>omotenashi-mind.com</dc:publisher>
-        <dc:title>はじめての電子書籍</dc:title>
-        <dc:language>ja</dc:language>
-        <meta property="dcterms:modified">2018-0pf9-19T13:50:40Z</meta>
-    </metadata>
-    <manifest>
-        <item id="nav" href="toc.xhtml" media-type="application/xhtml+xml" properties="nav" />
-        <item id="main_xhtml" href="main.xhtml" media-type="application/xhtml+xml" />
-    </manifest>
-    <spine page-progression-direction="default">
-        <itemref idref="nav" />
-        <itemref idref="main_xhtml" />
-    </spine>
-</package>
